# ledger/api/openapi.yaml
openapi: 3.0.3
info:
  title: Ledger API
  version: 0.1.0
  description: HTTP API for the ledger service.
paths:
...
  /ledgers/{slug}/accounts:
    post:
      operationId: createPayableAccount
      summary: Create a payable account in a ledger
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/LedgerSlug"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePayableAccountRequest"
      responses:
        "200":
          description: An identical request was previously completed; the existing account is returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatePayableAccountResponse"
        "201":
          description: The payable account was created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatePayableAccountResponse"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "404":
          description: The requested ledger does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                message: Ledger not found.
                error:
                  code: ledger_not_found
        "409":
          description: The request conflicts with the ledger's current state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                ledgerClosed:
                  summary: The ledger is closed
                  value:
                    message: Ledger is closed.
                    error:
                      code: ledger_closed
                holderConflict:
                  summary: The holder identity conflicts with an existing holder
                  value:
                    message: Holder information conflicts with an existing holder.
                    error:
                      code: holder_conflict
        "500":
          description: The server could not complete the request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                message: The server could not complete the request.
                error:
                  code: internal_server_error
  /ledgers/{slug}/accounts/{account_reference}:
    get:
      operationId: getAccount
      summary: Get a payable account and its balance
      description: Retrieves one payable account and exposes both its derived available amount and its underlying ledger counters. This API does not expose internal cash or revenue accounts.
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/LedgerSlug"
        - name: account_reference
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/AccountReference"
      responses:
        "200":
          description: The payable account was found and returned with its balances.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetAccountResponse"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "404":
          description: The requested ledger or payable account does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                ledgerNotFound:
                  summary: The ledger does not exist
                  value:
                    message: Ledger not found.
                    error:
                      code: ledger_not_found
                accountNotFound:
                  summary: The payable account does not exist in the ledger
                  value:
                    message: Account not found.
                    error:
                      code: account_not_found
        "500":
          description: The server could not complete the request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                message: The server could not complete the request.
                error:
                  code: internal_server_error
  /ledgers/{slug}/accounts/{account_reference}/statement:
    get:
      operationId: getAccountStatement
      summary: Get a payable account statement
      description: |
        Retrieves the journal movements affecting one payable account within a recorded-time period in deterministic chronological order (`recorded_at, journal entry identity, line_number`).

        ### Example movements

        | Recorded at | Journal reference | Kind | Direction | Amount | Balance |
        | --- | --- | --- | --- | ---: | ---: |
        | 2026-09-10 09:15 | `jrn_01M20J1QD2XB8K7G4N9CVF6T3A` | payment | Credit | 9,800 | 9,800 |
        | 2026-09-10 09:15 | `jrn_01M20J1QD2XB8K7G4N9CVF6T3A` | payment | Debit | 200 | 9,600 |
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/LedgerSlug"
        - name: account_reference
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/AccountReference"
        - name: from
          in: query
          required: false
          description: Inclusive recorded-time boundary.
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          required: false
          description: Exclusive recorded-time boundary.
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          required: false
          description: Bounded number of movements to return.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: cursor
          in: query
          required: false
          description: Opaque token returned by a preceding request.
          schema:
            type: string
      responses:
        "200":
          description: The payable account statement was successfully retrieved.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetAccountStatementResponse"
              example:
                account:
                  reference: acct_01M20H8704F1FDM1CFWSZDVJPV
                  holder_reference: hld_01M20H7XK4A9Q2TZ8VG6BFP31R
                  name: Acme Ltd
                period:
                  from: "2026-09-01T00:00:00Z"
                  to: "2026-10-01T00:00:00Z"
                opening_balance: 0
                closing_balance: 9600
                entries:
                  - journal_reference: jrn_01M20J1QD2XB8K7G4N9CVF6T3A
                    line_number: 1
                    kind: payment
                    direction: credit
                    amount: 9800
                    balance_after: 9800
                    description: Payment received
                    recorded_at: "2026-09-10T09:15:00Z"
                  - journal_reference: jrn_01M20J1QD2XB8K7G4N9CVF6T3A
                    line_number: 2
                    kind: payment
                    direction: debit
                    amount: 200
                    balance_after: 9600
                    description: Payment received
                    recorded_at: "2026-09-10T09:15:00Z"
                page:
                  limit: 50
                  previous_cursor: null
                  next_cursor: "opaque_next_cursor"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "404":
          description: The requested ledger or payable account does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                ledgerNotFound:
                  summary: The ledger does not exist
                  value:
                    message: Ledger not found.
                    error:
                      code: ledger_not_found
                accountNotFound:
                  summary: The payable account does not exist in the ledger
                  value:
                    message: Account not found.
                    error:
                      code: account_not_found
        "500":
          description: The server could not complete the request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                message: The server could not complete the request.
                error:
                  code: internal_server_error
  /ledgers/{slug}/entries:
    post:
      operationId: postJournalEntry
      summary: Post a journal entry to a ledger
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/LedgerSlug"
        - name: Idempotency-Key
          in: header
          required: true
          description: Caller-generated key that uniquely identifies this posting within the ledger.
          schema:
            $ref: "#/components/schemas/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostJournalEntryRequest"
      responses:
        "200":
          description: An identical request was previously completed; the existing entry is returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostJournalEntryResponse"
        "201":
          description: The journal entry was posted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostJournalEntryResponse"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "404":
          description: The ledger or one of the referenced accounts does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                ledgerNotFound:
                  summary: The ledger does not exist
                  value:
                    message: Ledger not found.
                    error:
                      code: ledger_not_found
                accountNotFound:
                  summary: A referenced account does not exist in the ledger
                  value:
                    message: Account not found.
                    error:
                      code: account_not_found
        "409":
          description: The posting conflicts with existing ledger state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                ledgerClosed:
                  summary: The ledger is closed
                  value:
                    message: Ledger is closed.
                    error:
                      code: ledger_closed
                accountClosed:
                  summary: A referenced account is closed
                  value:
                    message: Account is closed.
                    error:
                      code: account_closed
                insufficientFunds:
                  summary: An account balance restriction would be exceeded
                  value:
                    message: Insufficient funds.
                    error:
                      code: insufficient_funds
                idempotencyConflict:
                  summary: The idempotency key was previously used for different content
                  value:
                    message: Idempotency key conflicts with a previous request.
                    error:
                      code: idempotency_conflict
        "500":
          description: The server could not complete the request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                message: The server could not complete the request.
                error:
                  code: internal_server_error
components:
  schemas:
    LedgerSlug:
      type: string
      description: Opaque identifier assigned to the ledger.
      minLength: 3
      maxLength: 63
      pattern: '^[a-z0-9]+(?:_[a-z0-9]+)*$'
      example: ngn_ng
      x-validation-messages:
        minLength: must contain at least 3 characters
        maxLength: must not exceed 63 characters
        pattern: must contain lowercase letters and numbers separated by single underscores
    HolderReference:
      type: string
      description: Stable opaque reference assigned to a holder by the ledger.
      minLength: 1
      maxLength: 64
      example: hld_01K33YVADP5Z8B0T3X2Q91C6RH
    AccountReference:
      type: string
      description: Stable opaque reference assigned to an account by the ledger.
      minLength: 1
      maxLength: 64
      example: acct_01K33YV8M82N9MXP4E7J6B1QWK
      x-validation-messages:
        minLength: must not be empty
        maxLength: must not exceed 64 characters
    AccountStatus:
      type: string
      description: Whether the payable account itself is active or closed.
      enum:
        - active
        - closed
      example: active
    PayableAccountKind:
      type: string
      enum:
        - payable
      example: payable
...
    IdempotencyKey:
      type: string
      description: Caller-generated identifier used to make a ledger operation idempotent.
      minLength: 1
      maxLength: 255
      pattern: '.*\S.*'
      example: payment_123
      x-validation-messages:
        type: must be a string
        minLength: must not be empty
        maxLength: must not exceed 255 characters
        pattern: must not be blank
...
    PostJournalEntryRequest:
      type: object
      required:
        - kind
        - lines
      properties:
        kind:
          $ref: "#/components/schemas/JournalEntryKind"
        description:
          type: string
          minLength: 1
          maxLength: 500
          pattern: '.*\S.*'
          x-validation-messages:
            type: must be a string
            minLength: must not be empty
            maxLength: must not exceed 500 characters
            pattern: must not be blank
        effective_at:
          type: string
          format: date-time
        lines:
          type: array
          minItems: 1
          maxItems: 100
          items:
            $ref: "#/components/schemas/PostJournalEntryLine"
          x-validation-messages:
            type: must be an array
            minItems: must contain at least one journal line
            maxItems: must not contain more than 100 journal lines
...
    CreatePayableAccountRequest:
      type: object
      required:
        - external_id
        - name
      properties:
        external_id:
          type: string
          minLength: 1
          maxLength: 255
          pattern: '.*\S.*'
          x-go-name: ExternalID
          x-validation-messages:
            type: must be a string
            minLength: must not be empty
            maxLength: must not exceed 255 characters
            pattern: must not be blank
        name:
          type: string
          minLength: 1
          maxLength: 255
          pattern: '.*\S.*'
          x-validation-messages:
            type: must be a string
            minLength: must not be empty
            maxLength: must not exceed 255 characters
            pattern: must not be blank
    CreatePayableAccountResponse:
      type: object
      additionalProperties: false
      required:
        - holder_reference
        - reference
        - ledger_slug
        - external_id
        - name
        - kind
        - status
        - available
        - balances
        - created_at
      properties:
        holder_reference:
          $ref: "#/components/schemas/HolderReference"
        reference:
          $ref: "#/components/schemas/AccountReference"
        ledger_slug:
          $ref: "#/components/schemas/LedgerSlug"
        external_id:
          type: string
          minLength: 1
          example: merchant_123
          x-go-name: ExternalID
        name:
          type: string
          minLength: 1
          example: Acme Ltd
        kind:
          $ref: "#/components/schemas/PayableAccountKind"
        status:
          $ref: "#/components/schemas/AccountStatus"
        available:
          $ref: "#/components/schemas/AvailableBalance"
        balances:
          $ref: "#/components/schemas/AccountBalances"
        created_at:
          type: string
          format: date-time
    GetAccountResponse:
      type: object
      additionalProperties: false
      required:
        - reference
        - holder_reference
        - ledger_slug
        - kind
        - name
        - status
        - available
        - balances
        - created_at
      properties:
        reference:
          $ref: "#/components/schemas/AccountReference"
        holder_reference:
          $ref: "#/components/schemas/HolderReference"
        ledger_slug:
          $ref: "#/components/schemas/LedgerSlug"
        kind:
          $ref: "#/components/schemas/PayableAccountKind"
        name:
          type: string
          minLength: 1
          example: Acme Ltd
        status:
          $ref: "#/components/schemas/AccountStatus"
        available:
          $ref: "#/components/schemas/AvailableBalance"
        balances:
          $ref: "#/components/schemas/AccountBalances"
        created_at:
          type: string
          format: date-time
      example:
        reference: acct_01K33YV8M82N9MXP4E7J6B1QWK
        holder_reference: hld_01K33YVADP5Z8B0T3X2Q91C6RH
        ledger_slug: ngn_ng
        kind: payable
        name: Acme Ltd
        status: active
        available: 9800
        balances:
          debits_pending: 200
          credits_pending: 0
          debits_posted: 0
          credits_posted: 10000
        created_at: "2026-09-08T10:30:00Z"
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - message
        - error
      properties:
        message:
          type: string
          minLength: 1
        error:
          $ref: "#/components/schemas/ErrorInfo"
...
    GetAccountStatementResponse:
      type: object
      additionalProperties: false
      required:
        - account
        - period
        - opening_balance
        - closing_balance
        - entries
        - page
      properties:
        account:
          $ref: "#/components/schemas/StatementAccount"
        period:
          $ref: "#/components/schemas/StatementPeriod"
        opening_balance:
          type: integer
          format: int64
          description: Posted balance before the requested period in minor units.
          example: 0
        closing_balance:
          type: integer
          format: int64
          description: Posted balance at the end of the requested period in minor units.
          example: 9600
        entries:
          type: array
          items:
            $ref: "#/components/schemas/StatementMovement"
        page:
          $ref: "#/components/schemas/StatementPage"
  responses:
    InvalidRequest:
      description: The request does not satisfy the API contract.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorResponse"